Introduce FunctorT and TraverseT.
In the service of being able to have tests that are both robust and
clear.

`TraverseT` (and `FunctorT`) is a generalization of
`Recursive with Corecursive` providing `traverse` (and `map`) for types
that can’t implement one of `Recursive` or `Corecursive` (EG, `Free` and `Cofree`).

Transformations have been moved from the `Recursive` type class to the
new type classes, and some existing algebras have been generalized from
`Fix` to `T[_[_]]`.